refactor:delete textfieldFile,fixed setting page folder, added locales for profile
