+++ pages/profile/[id].tsx
@@ -1,0 +1,1 @@
+export { default } from '@/views/profile/ui/ProfilePage'
