--- pages/profile/settings/index.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/views/profile/ui/SettingsPage'
