--- src/components/ui/textfieldFile/index.ts
@@ -1,1 +1,0 @@
-export * from './textFieldFile'
