--- src/components/ui/textfieldFile/textfield.module.scss
@@ -1,6 +1,0 @@
-.btn{
-
-}
-.input{
-  display: none;
-}
