+++ src/locales/en.ts
@@ -2,7 +2,7 @@ import { confirmEmailPage } from '@/locales/en/confirm-email-page'
 import { confirmLogoutPage } from '@/locales/en/confirm-logout-page'
 import { passwordRecoveryPage } from '@/locales/en/password-recovery-page'
 import { privacyPolicyPage } from '@/locales/en/privacy-policy-page'
-import { profileSettingPage } from '@/locales/en/profile-setting-page'
+import { profilePage } from '@/locales/en/profile-page'
 import { signInPage } from '@/locales/en/sign-in-page'
 import { signUpPage } from '@/locales/en/sign-up-page'
 import { termsOfServicePage } from '@/locales/en/terms-of-service-page'
@@ -29,7 +29,7 @@ export const en = {
   },
   passwordRecoveryPage: passwordRecoveryPage,
   privacyPolicyPage: privacyPolicyPage,
-  profileSettingPage: profileSettingPage,
+  profilePage: profilePage,
   signInPage: signInPage,
   signUpPage: signUpPage,
   termsOfServicePage: termsOfServicePage,
