+++ src/locales/en/profile-page.ts
@@ -1,0 +1,11 @@
+export const profilePage = {
+  addProfilePhoto: 'Add a Profile Photo',
+  closeButton: 'Close',
+  deletePhotoQuestion: 'Are you sure you want to delete the photo?',
+  deleteProfilePhoto: 'Delete Photo',
+  noButton: 'No',
+  profileSettingsBtn: 'Profile Settings',
+  savePhoto: 'Save',
+  selectPhoto: 'Select from Computer',
+  yesButton: 'Да',
+}
