--- src/locales/en/profile-setting-page.ts
@@ -1,11 +1,0 @@
-export const profileSettingPage = {
-  profilePhoto: {
-    addProfilePhoto: 'Add a Profile Photo',
-    deletePhotoQuestion: 'Are you sure you want to delete the photo?',
-    deleteProfilePhoto: 'Delete Photo',
-    noButton: 'No',
-    savePhoto: 'Save',
-    selectPhoto: 'Select from Computer',
-    yesButton: 'Да',
-  },
-}
