+++ src/locales/ru.ts
@@ -3,7 +3,7 @@ import { confirmEmailPage } from '@/locales/ru/confirm-email-page'
 import { confirmLogoutPage } from '@/locales/ru/confirm-logout-page'
 import { passwordRecoveryPage } from '@/locales/ru/password-recovery-page'
 import { privacyPolicyPage } from '@/locales/ru/privacy-policy-page'
-import { profileSettingPage } from '@/locales/ru/profile-setting-page'
+import { profilePage } from '@/locales/ru/profile-page'
 import { signInPage } from '@/locales/ru/sign-in-page'
 import { signUpPage } from '@/locales/ru/sign-up-page'
 import { termsOfServicePage } from '@/locales/ru/terms-of-service-page'
@@ -22,7 +22,7 @@ export const ru: LocaleType = {
   },
   passwordRecoveryPage: passwordRecoveryPage,
   privacyPolicyPage: privacyPolicyPage,
-  profileSettingPage: profileSettingPage,
+  profilePage: profilePage,
   signInPage: signInPage,
   signUpPage: signUpPage,
   termsOfServicePage: termsOfServicePage,
