+++ src/locales/ru/profile-page.ts
@@ -1,0 +1,11 @@
+export const profilePage = {
+  addProfilePhoto: 'Добавить фото профиля',
+  closeButton: 'Закрыть',
+  deletePhotoQuestion: 'Вы уверены, что хотите удалить фото?',
+  deleteProfilePhoto: 'Удалить фото',
+  noButton: 'Нет',
+  profileSettingsBtn: 'Настройки профиля',
+  savePhoto: 'Сохранить',
+  selectPhoto: 'Выберите фото',
+  yesButton: 'Да',
+}
