--- src/locales/ru/profile-setting-page.ts
@@ -1,11 +1,0 @@
-export const profileSettingPage = {
-  profilePhoto: {
-    addProfilePhoto: 'Добавить фото профиля',
-    deletePhotoQuestion: 'Вы уверены, что хотите удалить фото?',
-    deleteProfilePhoto: 'Удалить фото',
-    noButton: 'Нет',
-    savePhoto: 'Сохранить',
-    selectPhoto: 'Выберите фото',
-    yesButton: 'Да',
-  },
-}
