+++ src/shared/enums/paths.enum.ts
@@ -9,8 +9,8 @@ export enum Paths {
   passwordRecovery = '/auth/password-recovery',
   privacyPolicy = '/docs/privacy-policy',
   profile = '/profile',
-  profileSettings = '/profile/settings',
   search = '/search',
+  settings = 'settings',
   signUp = '/auth/sign-up',
   statistics = '/statistics',
   termsOfService = '/docs/terms-of-service',
