--- src/views/profile/ui/ProfilePhoto.module.scss
@@ -1,36 +1,0 @@
-.root{
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  align-items: center;
-  width: 100%;
-  margin-top: 24px;
-}
-.avatar{
-
-}
-.dialog{
-  max-width: 492px;
-  max-height: 564px;
-}
-.dialogHeader{
-  display: flex;
-  justify-content: space-between;
-}
-.dialogContent{
-  max-width: 492px;
-
-}
-.blankImage{
-  margin-top: 48px;
-  margin-bottom: 60px;
-}
-.dialogBody{
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.saveButton {
-  margin: 36px 0 12px auto;
-}
